+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [2022-07-03]
+
+### Added
+
+- Observer pattern
+
+
 ## [2022-07-02]
 
 ### Added